+++ src/main/java/br/com/swconsultoria/nfe/Status.java
@@ -100,7 +100,7 @@ class Status {
             }
 
         } catch (RemoteException | XMLStreamException | JAXBException e) {
-            throw new NfeException(e.getMessage());
+            throw new NfeException(e.getMessage(),e);
         }
     }