+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Notas de versão
 
-- Atualizado Objeto e XSD NT 2023.001 1.20
-- Adicionado Validacao XML Manual
+- Atualizado Java Certificado (Ver https://github.com/Samuel-Oliveira/Java_Certificado)
+- Atualizado Cacert (Sefaz Minas)
+- Adicionado ZoneId direto a configuracao e removido da Chamada dos metodos. O padrao de ZoneId é America/Sao_Paulo